order

+++ migrations/20220217213100-create-order-item.js
@@ -10,7 +10,7 @@ module.exports = {
       },
       buyPrice: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       sellPrice: {
         allowNull: false,
